si0852 feat 패키지 이동

+++ src/main/java/com/hhplus/concert_ticketing/presentation/dto/response/BalanceResponse.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 @Getter
 public class BalanceResponse {
 
-    Long totalBalance;
+    Double totalBalance;
 
     @Builder
-    public BalanceResponse(Long totalBalance) {
+    public BalanceResponse(Double totalBalance) {
         this.totalBalance = totalBalance;
     }
 }
